Instruct usage of tags for installation and upgrade
Both the v1 and v2 instructions now ask the user to checkout a specific
tag when performing an installation, rathern than the master branch.
This will avoid any trouble in the future for new installations of
SendPortal.

The upgrade guide for v2 now instructs users to checkout the v2.0.0 tag.

Additionally, a new section has been added telling users who want to
stick with v1 to ensure they are explicitly using a v1 tag, as would be
expected for a new installation. This will help to avoid any trouble for
users who want to stick with v1, as master will stop being applicable to
them.